server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.4
pydantic[email]==2.10.4
python-dotenv==1.0.1
pymongo==4.9.2